Write dependent MV links to Metastore base table metadata
Keep a list of dependency materialized view names as a table parameter
of every base table. When a new MV is created, append its name to the lists.

Add and implement alterTable API in SemiTransactionalHiveMetastore.
With this commit, we only allow parameters modification for an existing table,
and will not touch table location, stats, etc.

Add alterTable API to ExtendedHiveMetastore and subclasses, and again
with this commit we only allow parameters modification for an existing table,
and will not touch table location, stats, etc. It calls alterTable() that is already implemented.